Fix a crash when the remote side stops and restarts video.
The video sink was added twice, and miss one refcount,
leading to disposal.

There now seems to be an occasional deadlock when ending
a call after video was stopped and restarted, but this is
better than a certain crash.

Also, after restarting video, there's a timing problem
which causes incoming video to be delayed.