Possible fix for a deadlock on hangup
A rare deadlock occurs when the QML sink delegate queues a delayed
stop via invokeMethod, and FarstreamChannel gets deleted just after
that, before stop actually gets called.
The QML code will wait for the actual stop to be performed in a
separate thread before continuing. The calling code may be buffer
chaining, via set_caps, and the sink's sink pad.
The FarstreamChannel deletion will deactivate the QML sink pad,
which waits for any buffer currently being streaming has gone before
going on. This never happens, as stop is never called.
The bit which I am not sure about is:
Both stop-via-invokeMethod and the FarstreamChannel deletion happen
via Qt events, and I think the stop event follows the deletion one,
so the stop waits in the event queue for the deletion to end, but
the deletion waits for the pad to be free, which waits for stop to
be performed.

This patch hopes to fix this by stopping the pipeline by setting it
to NULL before scheduling the deletion. This should ensure that all
streaming has ceased by the time the deletion gets scheduling, and
thus ensuring that the pad deactivation will not block.

Hopefully.