Fix audio often not starting at once when in a video call
When the audio source pad is added, the output bin's parent is
not always going to PLAYING yet, so syncing state with it will
only change the bin's state to PAUSED, letting it go to PLAYING
later.
Additionally, if the audio sink is async, it will forget its
pending state was PLAYING when going asynchronously to PAUSED,
and will then also go to PLAYING later.