Fix invisible video window while in a video call
If one starts a video call, but quickly pressed the end call button
as it appears, the call can be ended before the notification of the
call channel being created. The video window's opacity will follow
the UI logic and be hidden following the button pressed, but when
the channel is created, the video call will actually connect, and
the remote party will see video and hear audio, but the local party
will not see the video window.
Note that when this happens, the pending channel request is cancelled
but as it is asynchronous there is still some time window for the
channel to be made available between cancellation request and the
cancellation actually cancelling.